requirements.txt
requests
python-dotenv
spotipy
pandas
matplotlib
seaborn